requirements.txt
django>1.11
django-debug-toolbar
django-cors-headers
MySQL-python
django-directmessages
bcrypt
Pillow
